fix(studioPage): 네비게이션 라우트에 따른 텍스트 색상 변경 로직 수정
서브 메뉴의 라우트이면서 서브메뉴 버튼들이 접혀있을때 상위 메뉴버튼의 색상이 blue로 변경된다.
서브메뉴 버튼들이 펼쳐져있을때는 서브 메뉴의 버튼만 blue로 지정한다.

Resolves: #45
