+++ src/app/(route)/studio/[uid]/_components/Navigation/SubMenus.tsx
@@ -4,38 +4,11 @@ import React from 'react';
 import MenuButtonBox from '../common/MenuButtonBox.client';
 import NavDeptIcon from '@public/studioPage/NavDept.svg';
 import { useRouter } from 'next/navigation';
-
-interface MenuType {
-  text: string;
-  route: string;
-  isImplement: boolean;
-}
+import { getMenuList } from '@/app/_utils/studio/getMenuList';
 
 const SubMenu = ({ menuName, uid }: { menuName: string; uid: string }) => {
   const router = useRouter();
-
-  const menuList: Record<string, MenuType[]> = {
-    대시보드: [],
-    '방송 관리': [
-      { text: '방송하기', route: `studio/${uid}/live`, isImplement: true },
-      { text: '설정', route: `studio/${uid}/settings`, isImplement: false },
-      { text: '알림', route: `studio/${uid}/notice`, isImplement: false },
-      {
-        text: '리허설 방송 하기',
-        route: `studio/${uid}/rehearsal`,
-        isImplement: false,
-      },
-    ],
-    '시청자 관리': [
-      { text: '팔로워', route: `studio/${uid}/follower`, isImplement: true },
-      { text: '구독자', route: `studio/${uid}/subscriber`, isImplement: false },
-      {
-        text: '활동 제한',
-        route: `studio/${uid}/blocklist`,
-        isImplement: false,
-      },
-    ],
-  };
+  const menuList = getMenuList(uid);
 
   return (
     <div className="font-blackHanSans">
@@ -55,7 +28,7 @@ const SubMenu = ({ menuName, uid }: { menuName: string; uid: string }) => {
           gap={6}
           fontSize={14}
           onClick={() => {
-            if (item.isImplement) router.push(`/${item.route}`);
+            if (item.isImplement) router.push(item.route);
             else alert('추후에 구현 예정입니다.');
           }}
         />
